backend/requirements.txt
openai
pillow
moviepy
ffmpeg
ffmpeg-python
numpy
textwrap3
flask
python-dotenv
google-cloud-storage
flask-cors
requests
Pillow
